Added Docker-Compose support
* Added Docker-Compose

Docker-Compose project setup (sql server container dependency), sql server check and data seed and workflow action updated

* Workflow dotnet build update

* Workflow update

* Workflow update #2

* Workflow update #3

* Docker-Compose.yml update

* Workflow update #4 (how did I not see this)

+++ FileSystemAPI/FileSystem.API/Infrastructure/DataFactory.cs
@@ -15,7 +15,7 @@ namespace FileSystem.Infrastructure
 			var context = scope.ServiceProvider.GetRequiredService<FileSystemDbContext>();
 
 			var creatorService = context.Database.GetService<IRelationalDatabaseCreator>();
-			if (args.Length == 4 || !await creatorService?.ExistsAsync()!)
+			if ((args.Length == 4 || !await creatorService?.ExistsAsync()!) && context.Database.IsSqlServer())
 			{
 				if (args.Length == 4)
 					await context.Database.EnsureDeletedAsync();
